+++ spec/brcobranca/retorno/cnab240/sicoob_spec.rb
@@ -2,7 +2,7 @@
 
 RSpec.describe Brcobranca::Retorno::Cnab240::Sicoob do
   before do
-    @arquivo = File.join(File.dirname(__FILE__), '..', '..', '..', 'arquivos', 'CNAB240SICOOB.RET')
+    @arquivo = File.join(File.dirname(__FILE__), '..', '..', '..', 'fixtures', 'retorno', 'CNAB240SICOOB.RET')
   end
 
   it 'Transforma arquivo de retorno em objetos de retorno retornando somente as linhas de pagamentos de títulos sem registro' do